- Added ACL user.delete - small fix for calculation of final price in the product edit view, if the product belongs not to the main vendor - typo in sofort orderdone layout fixed - installer script, we allow now also the db type Jdiction_mysqli
git-svn-id: https://dev.virtuemart.net/svn/virtuemart@9809 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ branches/com_virtuemart.3.0.12.4/administrator/components/com_virtuemart/views/user/view.html.php
@@ -217,7 +217,11 @@ class VirtuemartViewUser extends VmViewAdmin {
 			JToolBarHelper::custom('toggle.user_is_vendor.1', 'publish','','COM_VIRTUEMART_USER_ISVENDOR');
 			JToolBarHelper::custom('toggle.user_is_vendor.0', 'unpublish','','COM_VIRTUEMART_USER_ISNOTVENDOR');
 			JToolBarHelper::divider();
-			JToolBarHelper::deleteList();
+
+			if (vmAccess::manager('user.delete')) {
+				JToolBarHelper::deleteList();
+			}
+
 			JToolBarHelper::editList();
 			self::showACLPref('user');
 			//This is intentionally, creating new user via BE is buggy and can be done by joomla
